All file changes including additions and removals though the fixing of the broken auto test is not present
