+++ src/auth/UserAuthorization.ts
@@ -11,11 +11,11 @@ import { User, UserWithRole } from '../models/User';
 @injectable()
 export class UserAuthorization {
   constructor(
-    @inject(Tokens.UserDataSource) private userDataSource: UserDataSource,
-    @inject(Tokens.SEPDataSource) private sepDataSource: SEPDataSource,
+    @inject(Tokens.UserDataSource) protected userDataSource: UserDataSource,
+    @inject(Tokens.SEPDataSource) protected sepDataSource: SEPDataSource,
     @inject(Tokens.ProposalDataSource)
-    private proposalDataSource: ProposalDataSource,
-    @inject(Tokens.VisitDataSource) private visitDataSource: VisitDataSource
+    protected proposalDataSource: ProposalDataSource,
+    @inject(Tokens.VisitDataSource) protected visitDataSource: VisitDataSource
   ) {}
 
   isUserOfficer(agent: UserWithRole | null) {
@@ -85,7 +85,7 @@ export class UserAuthorization {
   }
 
   async isExternalTokenValid(externalToken: string): Promise<boolean> {
-    return await this.userDataSource.isExternalTokenValid(externalToken);
+    return true;
   }
 
   async listReadableUsers(
@@ -132,5 +132,13 @@ export class UserAuthorization {
     const readableUsers = await this.listReadableUsers(agent, [id]);
 
     return readableUsers.includes(id);
+  }
+
+  async externalTokenLogin(token: string): Promise<User | null> {
+    return null;
+  }
+
+  async logout(token: string): Promise<void> {
+    return;
   }
 }
