+++ src/middlewares/graphql.ts
@@ -12,6 +12,7 @@ import { container } from 'tsyringe';
 import 'reflect-metadata';
 import { UserAuthorization } from '../auth/UserAuthorization';
 import baseContext from '../buildContext';
+import { Tokens } from '../config/Tokens';
 import { ResolverContext } from '../context';
 import { UserWithRole } from '../models/User';
 import federationSources from '../resolvers/federationSources';
@@ -74,7 +75,9 @@ const apolloServer = async (app: Express) => {
       let user = null;
       const userId = req.user?.user?.id as number;
       const accessTokenId = req.user?.accessTokenId;
-      const userAuthorization = container.resolve(UserAuthorization);
+      const userAuthorization = container.resolve<UserAuthorization>(
+        Tokens.UserAuthorization
+      );
 
       if (req.user) {
         if (accessTokenId) {
